saved-state error message 2/5 [clientLsp displays correct form]
Summary:
See previous diff in stack which explains the direction.

This diff makes clientIdeDaemon and clientIdeService expose the "short/medium/long" messages, and makes clientLsp.ml display them.

I decided that clientIdeService/clientIdeDaemon should maintain the "pure abstract" ontology
* short user message
* medium user message
* long user message
* debug details

And clientLsp should be the only one that turns that into concrete UI elements
* status-bar is "Hack: <short>"
* status-bar-tooltip is "<medium>. See Output>Hack for details."
* alert notification is "<medium>. See Output>Hack for details."
* window/logMessage is "<long> <link_to_debug>"
* Hh_logger.log is "<long> <debug_details>"

My idea is that the LSP adapter is the only one which knows that status-bar text needs to be prepended with the name of the language-service, and is the only one which knows where the <long> message will end up.

Note: the error messages currently look worse, as evinced by the changes to LSP integration tests. This is just a temporary work-in-progress. By the end of the diff stack they'll be all cleaned up again.

Reviewed By: arxanas

Differential Revision: D20480815

fbshipit-source-id: 128507aa25313aabfb8b503714903afdb6540675