Fix unwinding through identical frames
Summary:
It was possible for multiple frames in the same nesting to get the
same id, because we use the sp (which doesn't change through a given
nesting) and pc (which can be the same for multiple different
frames in the backtrace).

Add the fp value as the 'special' field in FrameId to differentiate.

Reviewed By: oulgen

Differential Revision: D20590529

fbshipit-source-id: 37293178d91d23b11a8eef16e2b352a24d67e334