Fix range-formatting issue
Summary:
In hackfmt's internal representation, opportunities for adding or removing line-breaks (called "splits") are controlled by behavioral switches (called "rules"). When hackfmt produces a candidate solution for adding/removing line breaks, it produces a mapping from rule IDs to bool, where true means "add a line break at all of the splits associated with this rule". The function modified in this diff generates such a "rule-bindings map" describing the way splits were broken (or not) in the original, unformatted source. This is so that when range-formatting, we can attempt to leave rules outside the range as they are, and add/remove line breaks only within the range.

When I originally wrote this function, I must have been interested in guaranteeing that we never fail to break any splits associated with an Always rule. This rule is used in situations where it is critically important that we add a line break (like after a single-line comment), but it's also used in some other situations where hackfmt unconditionally adds a line break (like after the opening curly brace of a function, or after the semicolon of a statement). By starting with binding all Always rules to break, I believe I was trying to ensure that hackfmt would never fail to add a line break after a single-line comment.

This shouldn't be necessary, though. When we actually USE this rule-bindings map, [we are careful to break the Always-rules which are within the formatting range](https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/hphp/hack/src/hackfmt/line_splitter.ml?commit=dae285b592d81882057e67ac0e1e231876462fd4&lines=99-102). On top of that, we can actually cause problems by requiring rules to break when they are actually not broken in the original source. In the added test case, hackfmt wants to add newlines after the opening curly brace of the lambda and before the closing curly brace, but it is outside of the formatting range. Because it wants to add a newline there, it believes it must also put each argument to `foo` on a new line, like this:

```
    foo(
      $a ==> {
        return $a;
      },
      $b,
    ),
```

That is to say, this formatting solution is considered to be invalid:

```
    foo($a ==> {
      return $a;
    }, $b),
```

Because we started by breaking the Always-rules, but do not propagate breakage (that is, binding the splits before each `foo` argument to be broken), we end up with an invalid rule-bindings map, which causes the solver to get stuck and refuse to explore the solve states which would cause us to correctly split up the `bar` call.

Reviewed By: arxanas

Differential Revision: D20572665

fbshipit-source-id: 40eef910400c018d4661d65ec5e9efb972ee9ceb