Cast JSON int as true iff =1
Summary:
Some HHVM options start off with an int value 2,
which indicates that certain functionality should be logged
but HHBC seemingly interprets it as `false` during the migration.

Eventually, logging typically disappears and option is passed as 0 or 1.

In JSON config passed to HackC, interpreting positive int values
other than 0 and 1, such as 2, as both:
- true (in code: `fun opts v -> ... v > 0`)
- false (in code: `fun opts v -> ... v = 1`)

means that the logic needs to *manually* specified for *every* option,
leading to a lot of boilerplate.

So, interpret bool-ish option (set as JSON ints) as:

  bool_val = (int_val = 1)

Reviewed By: dabek

Differential Revision: D20569996

fbshipit-source-id: 4412d7a664f444d53aae2628e9794088a09183e6