Use atomics to publish AllocDescs
Summary:
concurrent_vector::push_back updates the size of the vector but returns a reference to uninitialized memory for the element. This behavior causes a problem for RDS's forEachAlloc helpers, because they may read some uninitialized RDS handles.

Address this problem by keeping a separate atomic size for each vector. Only when we update the size do we "publish" new vector elements.

Reviewed By: markw65

Differential Revision: D20623620

fbshipit-source-id: f540550769ed68433b6a1c3c19a78b56b3cc61ca