Introduce pseudo-remote mode for remote check checking from hh_bulk_check
Summary:
This diff introduces a new `pseudoRemoteJobRunner` for remote type checking so that we can run remote type checking service on the same machine as the client.

This will make it much easier to compare the local and remote type checking, and shorten the testing/engineering cycle.

Currently, the only entrypoint is from `hh_bulk_check`.

Reviewed By: 2BitSalute

Differential Revision: D20413046

fbshipit-source-id: 2df511207968065ab2c802c1656a5655f97eb57f