Oxidize tast.ml
Summary: We need it if we want to pass Rust TAST back to OCaml. Slight drawback is that it will keep coming up in searches for `tast.rs` in addition to the "real" one. Maybe we should invest into oxidize script being able to rename files to make it less of a problem?

Reviewed By: CatherineGasnier

Differential Revision: D20527123

fbshipit-source-id: fa2dc009c2a736ad839ac72ea506b1b4ffbb8024