refactor clientIdeIncremental
Summary:
I'm on a mission to lift quarantine.

How it works today: you enter quarantine at the start of an operation you wish to do, and leave it at the end.
* SharedMem: during quarantine, all reads+writes go to a stack rather than sharedmem
* Local_memory caches: at start of quarantine it invalidates all caches related to ctx.entriess, and at the end of quarantine it invalidates the same caches again
* Names: it updates forward naming table at entry to quarantine so it reflects the quarantined files, either on the naming heap (sharedmem) or memory (localmem). It also updates the reverse naming table.

My immediate goal is I want to be able to remove entries for the context.

It feels like the cleanest way to think about ctx is this:
* Ctx has entries
* Ctx has caches. These caches are valid as of the entries.
* Ctx has naming-table stuff. This naming-table stuff is valid as of the entries.

The way I'd like to write ctx for localmem is:
* When you add an entry, then it invalidates caches and updates naming-table in the ctx
* When you remove an entry, then it invalidates caches and updates naming table
* Also, if we hear that a file on disk has changed, then we have to inform the ctx about this so it can update names accordingly.

That vision would let us do away with quarantine, at least for localmem backend. (sharedmem backend would still need it).

I don't know if I'll be able to reach that full vision. But I'd like to try. This diff is a tiny step in that direction. All it does is move "old_file_info = Naming_table.get_file_info" closer to the subsequent "update_naming_table".

Reviewed By: arxanas

Differential Revision: D20590001

fbshipit-source-id: a5b70a3c144d8533227fe358cceb89a1e3614153