Fix catch trace of ContEnter in Generator::send()
Summary:
ContEnter IR instruction previously teleported the already present value on the
stack to the generator's yield statement. This caused the default constructed
catch block to assume the value is on the stack and the new unwinder started to
issue an explicit erroneous decref of whatever was present on the stack. The
old unwinder was not affected, as unwinding of the generator body resulted in
correctly synced vmsp().

Fix this by first popping the value and then passing it explicitly to the IR
instruction.

Reviewed By: ricklavoie, oulgen

Differential Revision: D20634088

fbshipit-source-id: 16ea2e9ee0acfa58feec6ced66dba0761ce9f30b