add socket/bind/listen/accept/connect/fcntl, E*, F_, FD_ constants
Summary:
This is the next set of functionality that together let me write a unit test - fcntl is needed to set O_NONBLOCK, which is needed for server/client in the same request (via async); passing it to socket()/accept() is non-portable.

I've only implemented the trivial fcntl ops at present; as the types vary by op, they all need hardcoding.

Reviewed By: paulbiss

Differential Revision: D20546948

fbshipit-source-id: 3d3c8209d964924d5eb2d8bfd2888b9766dc5a85