Fix Dune build by adding `Core` to open-source build
Summary:
{D20370544} added a dependency on `Core`, which was present in Buck but not Dune.

I added `Core` to our Dune build, and upgrade `Core_kernel` while I was at it (to match the latest version of `Core`). The major changes are in `ocaml_deps_data.sh`, which upgrades the pinned package versions, and the generated `opam2-mini-repository.stamp` file, which gives instructions on how to regenerate the OPAM mini repo if you read it.

The rest of the changes are to fix `Core_kernel` upgrade issues. Most notably, `List.sort` now takes the labeled argument `~compare`, and `List.equal` no longer (?!) takes the labeled argument `~equal`.

Reviewed By: shiqicao

Differential Revision: D20608342

fbshipit-source-id: b2a54fd5451043bc6e3c033c12cfc92dffa01450