De-flake async tail-calls test
Summary:
Asio\join can elide up to one frame's worth of information, causing us to lose the function for that frame and the caller's file/line for the frame below it. This behavior is unrelated to the tail-call optimization (it happens with the release HHVM binary too) and is non-deterministic, but I haven't figured out exactly when it does or doesn't happen. Losing this information doesn't matter much in prod, because the root Asio\join call is so far up the call tree that no useful frames are elided this way.

This diff is a quick fix for places in the async tail-calls test affected by this behavior. We may lose the main() frame.

Reviewed By: ricklavoie

Differential Revision: D20653046

fbshipit-source-id: 996c431683707e27d8c649939007425d217ec879