Do not suppress formatting in files with manual section tags but no @partially-generated tag
Summary:
In {D20169730}, we began detecting manual section comments in files, and suppressing formatting outside of them, on the assumption that the file is partially generated, and code outside manual section comments must be generated. This is apparently not true--there are some files in WWW which contain manual section comments but are not partially generated (perhaps because they were generated in the past, but someone unsigned them or removed the header).

Instead of assuming a file is partially generated because it contains manual section comments, only suppress formatting outside manual sections when the file contains the text "partially-generated".

Reviewed By: arxanas

Differential Revision: D20656790

fbshipit-source-id: ab6c7b35dc9b54a81409bb69e5ab497756f63c63