Support retrieving function declarations
Summary: Instead of assuming every decl request is for a gconst, check the kind and return a gconst or function (whichever is requested), or panic otherwise. Implement Naming_provider APIs for the Decl_service backend using the decl-fetch APIs (in a future diff, we might want to extend the decl service protocol to support naming table requests, but this works for now).

Reviewed By: Wilfred

Differential Revision: D20352802

fbshipit-source-id: 416897e51cf94da2e634bc169e4e36f1bb954b94