Postpone localization of function parameters after type parameters
Summary:
At the moment, during a function call, type parameters are partially localized
(into type variables), then parameters are localized, and then type parameters
are finally fully processed (bounds are added to the type variables).

It means we don't process the function parameters with all the type information
we could. This diff tries to re-order these steps so we get the most of
type parameters.

Reviewed By: andrewjkennedy

Differential Revision: D20623010

fbshipit-source-id: 36def29be2c09ddf30c8906d205b0923bb3098f8