Make parameters conflict with live locals at function entry
Summary:
Local compaction did not take into account code that assumes locals are unset
at the start of a function.  Notably it might fold locals into unused parameter
slots.  Those parameters might not be unset at the start of the function.  This
could lead to bugs.

Make live locals at the start of a function all conflict with the parameters.
These locals (unless they themselves are parameters) depend on the slot being
unset at the start of the function.  If they are parameters they already were
disallowed from being remapped (as we need the slots there per our calling
convention).

Reviewed By: ricklavoie

Differential Revision: D20688842

fbshipit-source-id: c0dea12574853e30f3469de552f25928e210a8c6