Introduce dvarray type specializations
Summary:
To turn on dvarray type-hint notices in prod, we want to modify the JIT type system to understand these types, so that we only need to check a value once in a translation.

This diff adds this type-level support, in a way that's quite similar to the "vanilla" type specialization. We model "dvarray" as a single bit, and use that bit plus the kind to know that a value is a darray or a varray. This approach is better than using separate bits, because the rest of the HHVM also has some understanding of how array layouts change (e.g. that appending to a packed array doesn't change its layout) and we're adding more.

This diff also uses these types in a small number of places to optimize code when the type-hint notices are enabled. First off, whenever we do a packed or mixed layout check, if the notices are on, we instead do a dvarray check. This check buys us both the layout and the dvarray bit with a single 1-bit test. Second, we eliminate RaiseHackArrParamNotice / RaiseHackArrPropNotice if we know the array passes. Finally, we add a path to optimize FCallBuiltin Shapes::idx if the flag is on.

Reviewed By: jjgriego

Differential Revision: D20450850

fbshipit-source-id: 107e5db0f818ab23298b5326e6c194f4d0be0360