Change return type of load_dep_table_sqlite to void/unit
Summary:
In the next diff, I want to change the related function `hh_load_dep_table_blob` to return # of keys loaded instead of the # of milliseconds it takes to load the blob.

In order to make the API consistent, I'm changing `load_dep_table_sqlite` to return void instead of the # of milliseconds, since that is anyway a useless piece of info. We can more easily time the operation outside of the function, as we do with everything else.

Reviewed By: arxanas

Differential Revision: D19841074

fbshipit-source-id: 1de42958564d235723843938d9bde1c10868a110