Handle unprocessed predecessors in pdce-inlining
Summary:
The changes in D20646942 assumed that by processing the cfg using an rpo sort all blocks will be processed only after their predecessors. While this isn't the case in the presence of a cycle, we only depend on having processed predecessors to extends Jump instructions when creating the Phi.

Instead we insert `ctx.deadFp` into the the `Jump` in these cases so that it will be updated appropriately when the block is processed.

Reviewed By: ricklavoie

Differential Revision: D20732191

fbshipit-source-id: 39331dc269a354d683ad34e87ec76178398ca753