Don't bubble up errors we don't care about in the direct decl parser.
Summary: All that we really care about in method bodies is yields, so if we see a yield and an error in a list we can drop the error instead of bubbling it up.  This is important now that we're partially processing expressions, because we only want to support expressions that can occur in constant initializers, but having only partial expression support means that a lot of expressions inside method bodies are now errors.

Reviewed By: Wilfred

Differential Revision: D19907546

fbshipit-source-id: 0cb0c27e4eaa68b7cbe033b04e0e9a44d7d8c818