Handle Phi nodes covering TFramePtr in collapseDiamond
Summary: Code in memory effects (and elsewhere) depends on being able to trace through DefLabels and passthrough instructions to find a canonical definition for a TFramePtr, generating TFramePtr values using Select violates invariants in that logic. As I explain in the comments I don't see how it would actually be possible to end up with diamond control flow that both can be collapsed and needs to phi together differing values of TFramePtr.

Reviewed By: ricklavoie

Differential Revision: D20752331

fbshipit-source-id: 006cf756a460f14c9c881c08f568010ec4f3acde