Drop count requirement for iterator specialization
Summary: This change shouldn't have any significant perf effect - the count bound was already low enough that any site under the bound is too cold to really matter for perf. The main reason I want to land it is to make reading production TC dumps easier - it's confusing when an IterInit over, say, a Vec doesn't get specialized because of the count.

Reviewed By: ricklavoie

Differential Revision: D20724326

fbshipit-source-id: 2c45156bfcfb82a5d8d9a4cc65165e01c3cb8da1