Abstract away local accessing logic in dec-ref unique stubs
Summary:
We want to minimize the number of places where we're making
assumptions about the frame layout. The generic dec-ref helper stubs
is one of the places where we directly emit Vasm to access the
locals. Add helper functions to code-gen-helpers.cpp to centralize the
logic of obtaining pointers to local's types and values and use that
in the stubs instead.

Reviewed By: mxw, jano

Differential Revision: D20753292

fbshipit-source-id: 56328862643a9985889745b66e401b2eff07a6c7