Port emit_class_get/emit_special_function, fix emit_method, and add print implementations
Summary:
- Port `emit_class_get` and `HH\\fun` case in `emit_special_function`
- Fixed a condition check on memoize method in `emit_method`
- Fixed print for `Call` expression
- Implemented print functions for `new` expressions and class consts
- A problem:
when printing class consts we need to replace the `self` in something like `self::foo()` with the class name, which requires the state from the emitter. I'm thinking maybe we could the emitter in `Context`, but left a TODO for now. If this sounds good I will implement it in a separate diff

Reviewed By: shiqicao

Differential Revision: D20752489

fbshipit-source-id: 3c211bbd18ae5b20137a81fb7015f083ef8e0112