RFC rename peak_changed_files_queue_size
Summary:
I had a hard time understanding the meaning of `peak_changed_files_queue_size` -- because it doesn't record the peak size of the changed_files_queue!

Indeed, upon File_changed events when we add a new item to the changed_files_queue, we do
```
      let changed_files_denominator =
        initialized_state.changed_files_denominator + 1
```
That's definitely not measureing peak queue size. It's measuring something else.

What it's measuring is a user-facing concept "the size of the current batch of updates from the perspective of the user". And this concept has to cope with new files being added to the batch even as the batch is being processed.

I'm renaming it to `changed_files_denominator` because that makes clear that the sole role of this function is to be the denominator in the display of how many files have to be processed.

Reviewed By: arxanas

Differential Revision: D20627334

fbshipit-source-id: 513400559a840957202eb34c6c5b42044035cef0