Remove __halt_compiler
Summary:
`__halt_compiler` was a [PHP feature](https://www.php.net/manual/en/function.halt-compiler.php) that Hack supported while it maintained back compatibility with PHP. Currently, we don't support all of the associated functionality of `__halt_compiler`, such as the magic constant: `__COMPILER_HALT_OFFSET_`.

There are not many [good use cases](https://stackoverflow.com/questions/5270486/whats-halt-compiler-in-php-for) for such a feature and it complicates Hack's AST and means we have to provide special support in different parts of the pipeline for it. This diff removes support for it from Hack.

Reviewed By: fredemmott, dabek

Differential Revision: D20739868

fbshipit-source-id: bc39c64959eedba7cdeae1fdde4fac24d3869563