Simplify VariableUnserializer ref tracking
Summary:
VariableUnserializer supports "reference" elements such as "R:2" that refers back to earlier serialized values. We implement these references by tracking pointers to the old data, rather than the values themselves. This distinction affects behavior, because e.g. if we deserialize an array with multiple instances of a key, when we overwrite its value, we'll affect references to the old value.

Rather than eliminating these references completely, I opted to eliminate generic calls to "lvalForce" to make this code layout-safe. We won't support this generic operation going forward, but we will support an LvalForce method (with key) on dict-ish arrays and an LvalForceNew method (which appends) on vec-ish arrays. By modifying the code to call these layout-specific members, we can restrict the general interface while reducing the number of virtual calls we make.

In the process, I made a number of other simplifications to this code:
- To handle dynamic props, reuse the existing "makeDynProp" rather than redoing it.
- Replace the RefInfo struct with a simple tv_rval.
- Pull out the element termination check into "checkElemTermination".
- Consolidate the UnserializeModes into just three: Value, Key, and ColKey.

Reviewed By: ricklavoie

Differential Revision: D20773163

fbshipit-source-id: ac2938855f56023a7c42264cf07ce09acb880775