Fix run.php in --server and --cli-server modes.
Summary:
D16520159 changed hhvm_cmd_impl but only updated one of the call sites. Fix it by renaming the $test parameter to $autoload_db_prefix (to make it clear what it does) and allowing it to be null. Pass null in when starting a server process.

Also fix the code that attempted to set the server's thread count, which had bit-rotted in the meantime.

Reviewed By: fredemmott

Differential Revision: D20797644

fbshipit-source-id: aec70a6fda0878eb828d6927f9a36fab8951a63b