Fix infinite loop in is_enforceable
Summary:
Direct cycles in type definitions such as
```
type T = T;
```
or
```
type T = T2;
type T2 = T;
```
on their own correctly produce a "Cyclic typedef" error, but cause the type checker to go into an infinite loop if the type alias appears in a function signature e.g.
```
function foo(T $x):void { }
```
This is due to the lack of a "visited" check in `Typing_enforceability.is_enforceable`.

For future, it seems to me that a more maintainable solution would be to mark the type *definition* as bad, and thereby avoid repeated checks for cycles in other parts of the type checker.

Reviewed By: Wilfred

Differential Revision: D20763755

fbshipit-source-id: 56d9c55529b24760eb79ea92bfdfc909831433f3