Fix LSP integration tests for Python upgrade
Summary:
The Python upgrade in platform009 seems to change the line numbers in the call stack frame. It now seems to return the line of the first function in a function call chain under some circumstances, but I couldn't get a good repro.

I'm not sure if this is a bug in Python? But to unblock the platform009 upgrade, this commit updates the tests to use a non-chained function called `line()` that's passed into every `request`, which we then use to get the line number for error display. It's kind of ugly, but it works.

Reviewed By: ljw1004

Differential Revision: D20780754

fbshipit-source-id: 7b4dc8df060a22012b0e5c5a1a3bb4a8f719b9de