canon 1/3 get_from_sqlite needn't take name
Summary:
Trivial.

I found it confusing that the caller of the function (1) passed a name into get_and_cache, (2) received a name in it get_from_sqlite fallback.

What was the relation of the second name to the first?

In fact the answer is the two were always the same. So I removed the name in the callback to avoid redundancy. (this is in preparation for the following diff, which uses canon_keys, where I prefer to be clear that get_and_cache is doing nothing canonical).

Reviewed By: arxanas

Differential Revision: D20784432

fbshipit-source-id: 360bea8911f1a008676760c0aa01ce62dd546581