Refactor closure_convert to not use Def_inline nodes
Summary:
Currently, `Def_inline` is an AAST node that is only used in one place: closure_convert (for codegen/emission). However - we still have other logic in the Hack toolchain that still acts as if `Def_inline` might someday be relevant, such as parsing support (we don't support inline definitions to begin with). This diff attempts to refactor closure_convert such that we do not need to use `Def_inline`.

How the system works today:

closure_convert takes all toplevel class, record, constant and typedef declarations and adds a new stub copy of each def with the only relevant piece of information being that def's "count" (i.e. this is the 0th, 1st, 2nd class/typedef/constant defined in the file). Then, this information is used in emit_statement to emit a bytecode in main.

How the system is refactored:

closure_convert will add that piece of information onto the def itself if it is a top level def. (c_declaration, td_declaration, etc.) emit_defs in emit_body will then do the emission that emit_statement would have done.

Reviewed By: Wilfred

Differential Revision: D20752103

fbshipit-source-id: e78720b2f4920f41f386d24fbb3de6d33f1df5b1