Tempfile should have an API allowing to specify the base directory
Summary:
The library is making an assumption that it can freely write to /tmp/, but in some cases, the temporary directory could be elsewhere. Sometimes the system /tmp/ directory is not writeable, though usually that's an infrastructure issue, and at other times, you want to specify a parent directory in which to generate a new temp one for other reasons. Those reasons might include:
* there is a temp directory that will get cleaned based on a certain policy, such as the end of a job
* you want to create a random temp directory within a temp directory - you could then chain Tempfile.mkdtemp_with_dir calls.

This diff adds an API to allow specifying the base temp directory in which to create further temp directories.

Reviewed By: yinghuitan

Differential Revision: D20721898

fbshipit-source-id: c99054afc42e311dc0125d9136d8e50caf57f427