Replace use of bools etc in fun_type by bitwise flags
Summary:
Some of our OCaml type definitions are quite wasteful in space. This is particularly noticeable if they're used in the decl cache, where objects persist across major GC collections. A good source of low-hanging fruit is bools, which take up a full word each but can be represented using bitwise operations instead. (Even though OCaml does not support bitfields, the amount of code to manipulate bits required is small.)

This is the first of a stack of diffs that uses bit packing to reduce the size of our values. Here, function types, which are used for every function element and class elements that are methods.

Old size of `fun_type`: 16 words (128 bytes), including a nested tuple for `ft_tparams`.
New size of `fun_type`: 8 words (64 bytes).

Reviewed By: losvald

Differential Revision: D20791777

fbshipit-source-id: e39f3c04718cfba589f0e0519d59bc8c8548f041