Remove max arity in ft_arity, use List.length fp_params instead
Summary: Easy optimization: there is no need to store the max arity in the `ft_arity` field of function types, as this can be determined from `fp_params` instead. For normal arities, this reduces the space used for the arity (in every function type) from 3 words (24 bytes) down to 1 word (8 bytes).

Differential Revision: D20791863

fbshipit-source-id: 70acf7484926e30c12e8bcf39afe3b0338dcc3b8