Eliminate rval ArrayIter methods
Summary:
Rather than exposing the rval to the value at a position, ArrayIter should expose the TypedValue there instead. There are basically no call-sites that want to load the type and not the value in the common case (although many call-sites first load the type and then bail out if it's unexpected).

This diff is 95% automatic. The only nontrivial change is some code I wrote for array-provenance that mutates array values in place. For this specific use case, I replace the rval lookup with fast in-place access for packed and mixed arrays, and fall back to a vtable set otherwise. We have good test coverage of this site for packed and mixed arrays, and for COW and non-COW cases - see hphp/test/slow/array_provenance/recursive-provenance.php.

Reviewed By: ricklavoie

Differential Revision: D20811743

fbshipit-source-id: cff2b10b541739f994090b94ecc51d6cb8d3aea7