Clean up generate-build-info.sh
Summary:
I noticed generate-build-info.sh was sometimes taking a long time
(many minutes) on edenfs. While investigating better ways to do this
on eden, I noticed that we're inluding lots of unnecessary files in
the sha1sum, including 20000 from the hack/test directory.

This cleans up the script, and tightens the regex to exclude more
unwanted files, dropping us from over 25000 files to ~5000. I also
noticed that calling stat on a file thats not in the cache is very
slow in eden, so filter by name first, then check for symlinks,
instead of the other way around.

Its hard to measure the perf win, because it was only slow when the
eden cache was cold, and as soon as you run the script it warms the
cache.

Reviewed By: oulgen, alexeyt

Differential Revision: D20806227

fbshipit-source-id: 1fe34e4e361ae8e8749c59c5cd540cf3f72abfe1