Replace `self` in type constant accesses with current class name
Summary:
Naming replaces the token `self` in a typeconst access with the name of the containing class, with a position pointing to the class name at the top of the class declaration. In the direct decl smart constructors, we do not get access to the class name until we have finished parsing the class (in `make_classish_declaration`).

We could change the CST to get signal earlier by introducing a ClassishDeclarationHeader node (and it might be worthwhile to do so in the future). But we can work around the problem for now using the same technique we use for namespaces--we wait for the `class` token, then store the next Name we see as the name of the current class.

Reviewed By: Wilfred

Differential Revision: D20745037

fbshipit-source-id: 180b762bdd42a312a00ad604f8559ce907901a25